+++ docs/conf.py
@@ -47,7 +47,7 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = "DonkeyCar OpenAI Gym"
+project = "DonkeyCar Gymnasium"
 copyright = "2019, Leigh Johnson"
 author = "Leigh Johnson"
 
@@ -125,7 +125,7 @@ latex_elements = {
 # (source start file, target name, title, author, documentclass
 # [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, "gym_donkeycar.tex", "OpenAI Gym Environments for Donkey CarDocumentation", "Leigh Johnson", "manual")
+    (master_doc, "gym_donkeycar.tex", "Gymnasium Environments for Donkey Car Documentation", "Leigh Johnson", "manual")
 ]
 
 
@@ -133,7 +133,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [(master_doc, "gym_donkeycar", "OpenAI Gym Environments for Donkey CarDocumentation", [author], 1)]
+man_pages = [(master_doc, "gym_donkeycar", "Gymnasium Environments for Donkey Car Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------
@@ -145,7 +145,7 @@ texinfo_documents = [
     (
         master_doc,
         "gym_donkeycar",
-        "OpenAI Gym Environments for Donkey CarDocumentation",
+        "Gymnasium Environments for Donkey Car Documentation",
         author,
         "gym_donkeycar",
         "One line description of project.",