+++ gym_donkeycar/core/client.py
@@ -9,6 +9,7 @@ asynchronous manner.
 
 Author: Tawn Kramer
 """
+
 import json
 import logging
 import select